修复一个弹幕id过长的问题

+++ src/BaseDanmaku.hpp
@@ -94,7 +94,7 @@ namespace WTFDanmaku {
             return mTimestamp;
         }
 
-        inline uint32_t GetDanmakuId() {
+        inline long long GetDanmakuId() {
             return mDanmakuId;
         }
     protected:
@@ -111,7 +111,7 @@ namespace WTFDanmaku {
         uint32_t mTextShadowColor = 0;
         uint32_t mFilterParams = 0;
         time_t mTimestamp = 0;
-        uint32_t mDanmakuId = 0;
+        long long mDanmakuId = 0;
         int mMeasureFlag = 0;
         shared_ptr<Renderable> mRenderable;
     };
