+++ src/DanmakuFactory.cpp
@@ -6,7 +6,7 @@
 
 namespace WTFDanmaku {
 
-    DanmakuRef DanmakuFactory::CreateDanmaku(DanmakuType type, time_t time, std::wstring& comment, int fontSize, int fontColor, time_t timestamp, int danmakuId) {
+    DanmakuRef DanmakuFactory::CreateDanmaku(DanmakuType type, time_t time, std::wstring& comment, int fontSize, int fontColor, time_t timestamp, long long danmakuId) {
         DanmakuRef danmaku = nullptr;
 
         switch (type) {
